FIX: 修复Ubuntu上编译enb和ue的问题

+++ lib/src/phy/rf/rf_blade_imp.c
@@ -38,17 +38,17 @@
 #define CONVERT_BUFFER_SIZE 240*1024
 
 typedef struct {
-  struct bladerf *dev; 
+  struct bladerf *dev;
   uint32_t rx_rate;
   uint32_t tx_rate;
-  int16_t rx_buffer[CONVERT_BUFFER_SIZE]; 
-  int16_t tx_buffer[CONVERT_BUFFER_SIZE]; 
-  bool rx_stream_enabled; 
-  bool tx_stream_enabled; 
+  int16_t rx_buffer[CONVERT_BUFFER_SIZE];
+  int16_t tx_buffer[CONVERT_BUFFER_SIZE];
+  bool rx_stream_enabled;
+  bool tx_stream_enabled;
   srslte_rf_info_t info;
 } rf_blade_handler_t;
 
-srslte_rf_error_handler_t blade_error_handler = NULL; 
+srslte_rf_error_handler_t blade_error_handler = NULL;
 
 void rf_blade_suppress_stdout(void *h) {
   bladerf_log_set_verbosity(BLADERF_LOG_LEVEL_SILENT);
@@ -62,16 +62,16 @@ void rf_blade_register_error_handler(void *notused, srslte_rf_error_handler_t ne
 bool rf_blade_rx_wait_lo_locked(void *h)
 {
   usleep(1000);
-  return true; 
+  return true;
 }
 
 const unsigned int num_buffers    = 256;
-const unsigned int ms_buffer_size_rx = 1024; 
-const unsigned int buffer_size_tx = 1024;  
+const unsigned int ms_buffer_size_rx = 1024;
+const unsigned int buffer_size_tx = 1024;
 const unsigned int num_transfers  = 32;
 const unsigned int timeout_ms     = 4000;
 
- 
+
 char* rf_blade_devname(void* h)
 {
   return DEVNAME;
@@ -79,9 +79,9 @@ char* rf_blade_devname(void* h)
 
 int rf_blade_start_tx_stream(void *h)
 {
-  int status; 
+  int status;
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
-  
+
   status = bladerf_sync_config(handler->dev,
                                 BLADERF_MODULE_TX,
                                 BLADERF_FORMAT_SC16_Q11_META,
@@ -91,26 +91,26 @@ int rf_blade_start_tx_stream(void *h)
                                 timeout_ms);
   if (status != 0) {
     fprintf(stderr, "Failed to configure TX sync interface: %s\n", bladerf_strerror(status));
-    return status; 
+    return status;
   }
   status = bladerf_enable_module(handler->dev, BLADERF_MODULE_TX, true);
   if (status != 0) {
     fprintf(stderr, "Failed to enable TX module: %s\n", bladerf_strerror(status));
     return status;
   }
-  handler->tx_stream_enabled = true; 
+  handler->tx_stream_enabled = true;
   return 0;
 }
 
 int rf_blade_start_rx_stream(void *h, bool now)
 {
-  int status; 
+  int status;
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
-    
+
   /* Configure the device's RX module for use with the sync interface.
      * SC16 Q11 samples *with* metadata are used. */
-  uint32_t buffer_size_rx = ms_buffer_size_rx*(handler->rx_rate/1000/1024); 
-  
+  uint32_t buffer_size_rx = ms_buffer_size_rx*(handler->rx_rate/1000/1024);
+
   status = bladerf_sync_config(handler->dev,
                                 BLADERF_MODULE_RX,
                                 BLADERF_FORMAT_SC16_Q11_META,
@@ -131,7 +131,7 @@ int rf_blade_start_rx_stream(void *h, bool now)
                                 timeout_ms);
   if (status != 0) {
     fprintf(stderr, "Failed to configure TX sync interface: %s\n", bladerf_strerror(status));
-    return status; 
+    return status;
   }
   status = bladerf_enable_module(handler->dev, BLADERF_MODULE_RX, true);
   if (status != 0) {
@@ -143,7 +143,7 @@ int rf_blade_start_rx_stream(void *h, bool now)
     fprintf(stderr, "Failed to enable TX module: %s\n", bladerf_strerror(status));
     return status;
   }
-  handler->rx_stream_enabled = true; 
+  handler->rx_stream_enabled = true;
   return 0;
 }
 
@@ -161,7 +161,7 @@ int rf_blade_stop_rx_stream(void *h)
     return status;
   }
   handler->rx_stream_enabled = false;
-  handler->tx_stream_enabled = false; 
+  handler->tx_stream_enabled = false;
   return 0;
 }
 
@@ -169,31 +169,31 @@ void rf_blade_flush_buffer(void *h)
 {
 }
 
-bool rf_blade_has_rssi(void *h) 
+bool rf_blade_has_rssi(void *h)
 {
-  return false; 
+  return false;
 }
 
-float rf_blade_get_rssi(void *h) 
+float rf_blade_get_rssi(void *h)
 {
   return 0;
 }
 
 int rf_blade_open_multi(char *args, void **h, uint32_t nof_channels)
 {
-  return rf_blade_open(args, h); 
+  return rf_blade_open(args, h);
 }
 
 int rf_blade_open(char *args, void **h)
 {
-  *h = NULL; 
-  
+  *h = NULL;
+
   rf_blade_handler_t *handler = (rf_blade_handler_t*) malloc(sizeof(rf_blade_handler_t));
   if (!handler) {
     perror("malloc");
-    return -1; 
+    return -1;
   }
-  *h = handler; 
+  *h = handler;
 
   printf("Opening bladeRF...\n");
   int status = bladerf_open(&handler->dev, args);
@@ -201,9 +201,9 @@ int rf_blade_open(char *args, void **h)
     fprintf(stderr, "Unable to open device: %s\n", bladerf_strerror(status));
     return status;
   }
-  
+
   //bladerf_log_set_verbosity(BLADERF_LOG_LEVEL_VERBOSE);
-  
+
   /* Configure the gains of the RX LNA and RX VGA1*/
   status = bladerf_set_lna_gain(handler->dev, BLADERF_LNA_GAIN_MAX);
   if (status != 0) {
@@ -220,7 +220,7 @@ int rf_blade_open(char *args, void **h)
     fprintf(stderr, "Failed to set TX VGA1 gain: %s\n", bladerf_strerror(status));
     return status;
   }
-  handler->rx_stream_enabled = false; 
+  handler->rx_stream_enabled = false;
   handler->tx_stream_enabled = false;
 
   /* Set info structure */
@@ -240,13 +240,13 @@ int rf_blade_close(void *h)
   return 0;
 }
 
-void rf_blade_set_master_clock_rate(void *h, double rate) 
+void rf_blade_set_master_clock_rate(void *h, double rate)
 {
 }
 
-bool rf_blade_is_master_clock_dynamic(void *h) 
+bool rf_blade_is_master_clock_dynamic(void *h)
 {
-  return true; 
+  return true;
 }
 
 double rf_blade_set_rx_srate(void *h, double freq)
@@ -258,7 +258,7 @@ double rf_blade_set_rx_srate(void *h, double freq)
     fprintf(stderr, "Failed to set samplerate = %u: %s\n", (uint32_t) freq, bladerf_strerror(status));
     return -1;
   }
-  if (handler->rx_rate < 2000000) { 
+  if (handler->rx_rate < 2000000) {
     status = bladerf_set_bandwidth(handler->dev, BLADERF_MODULE_RX, handler->rx_rate, &bw);
     if (status != 0) {
       fprintf(stderr, "Failed to set bandwidth = %u: %s\n", handler->rx_rate, bladerf_strerror(status));
@@ -283,7 +283,7 @@ double rf_blade_set_tx_srate(void *h, double freq)
   if (status != 0) {
     fprintf(stderr, "Failed to set samplerate = %u: %s\n", (uint32_t) freq, bladerf_strerror(status));
     return -1;
-  }  
+  }
   status = bladerf_set_bandwidth(handler->dev, BLADERF_MODULE_TX, handler->tx_rate, &bw);
   if (status != 0) {
     fprintf(stderr, "Failed to set bandwidth = %u: %s\n", handler->tx_rate, bladerf_strerror(status));
@@ -294,8 +294,8 @@ double rf_blade_set_tx_srate(void *h, double freq)
 
 double rf_blade_set_rx_gain(void *h, double gain)
 {
-  int status; 
-  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;  
+  int status;
+  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   status = bladerf_set_rxvga2(handler->dev, (int) gain);
   if (status != 0) {
     fprintf(stderr, "Failed to set RX VGA2 gain: %s\n", bladerf_strerror(status));
@@ -306,8 +306,8 @@ double rf_blade_set_rx_gain(void *h, double gain)
 
 double rf_blade_set_tx_gain(void *h, double gain)
 {
-  int status; 
-  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;  
+  int status;
+  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   status = bladerf_set_txvga2(handler->dev, (int) gain);
   if (status != 0) {
     fprintf(stderr, "Failed to set TX VGA2 gain: %s\n", bladerf_strerror(status));
@@ -318,9 +318,9 @@ double rf_blade_set_tx_gain(void *h, double gain)
 
 double rf_blade_get_rx_gain(void *h)
 {
-  int status; 
-  int gain; 
-  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;  
+  int status;
+  int gain;
+  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   status = bladerf_get_rxvga2(handler->dev, &gain);
   if (status != 0) {
     fprintf(stderr, "Failed to get RX VGA2 gain: %s\n",
@@ -332,9 +332,9 @@ double rf_blade_get_rx_gain(void *h)
 
 double rf_blade_get_tx_gain(void *h)
 {
-  int status; 
-  int gain; 
-  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;  
+  int status;
+  int gain;
+  rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   status = bladerf_get_txvga2(handler->dev, &gain);
   if (status != 0) {
     fprintf(stderr, "Failed to get TX VGA2 gain: %s\n",
@@ -360,34 +360,34 @@ srslte_rf_info_t *rf_blade_get_info(void *h)
 double rf_blade_set_rx_freq(void *h, double freq)
 {
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
-  uint32_t f_int = (uint32_t) round(freq);
+  uint64_t f_int = (uint64_t) round(freq);
   int status = bladerf_set_frequency(handler->dev, BLADERF_MODULE_RX, f_int);
   if (status != 0) {
-    fprintf(stderr, "Failed to set samplerate = %u: %s\n",
-            (uint32_t) freq, bladerf_strerror(status));
+    fprintf(stderr, "Failed to set samplerate = %lu: %s\n",
+            (uint64_t) freq, bladerf_strerror(status));
     return -1;
   }
   f_int=0;
   bladerf_get_frequency(handler->dev, BLADERF_MODULE_RX, &f_int);
-  printf("set RX frequency to %u\n", f_int);
-  
+  printf("set RX frequency to %lu\n", f_int);
+
   return freq;
 }
 
 double rf_blade_set_tx_freq(void *h, double freq)
 {
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
-  uint32_t f_int = (uint32_t) round(freq);
+  uint64_t f_int = (uint64_t) round(freq);
   int status = bladerf_set_frequency(handler->dev, BLADERF_MODULE_TX, f_int);
   if (status != 0) {
-    fprintf(stderr, "Failed to set samplerate = %u: %s\n",
-            (uint32_t) freq, bladerf_strerror(status));
+    fprintf(stderr, "Failed to set samplerate = %lu: %s\n",
+            (uint64_t) freq, bladerf_strerror(status));
     return -1;
   }
-  
+
   f_int=0;
   bladerf_get_frequency(handler->dev, BLADERF_MODULE_TX, &f_int);
-  printf("set TX frequency to %u\n", f_int);
+  printf("set TX frequency to %lu\n", f_int);
   return freq;
 }
 
@@ -396,7 +396,7 @@ void rf_blade_set_tx_cal(void *h, srslte_rf_cal_t *cal) {
   bladerf_set_correction(handler->dev, BLADERF_MODULE_TX, BLADERF_CORR_FPGA_PHASE, cal->dc_gain);
   bladerf_set_correction(handler->dev, BLADERF_MODULE_TX, BLADERF_CORR_FPGA_GAIN, cal->dc_phase);
   bladerf_set_correction(handler->dev, BLADERF_MODULE_TX, BLADERF_CORR_LMS_DCOFF_I, cal->iq_i);
-  bladerf_set_correction(handler->dev, BLADERF_MODULE_TX, BLADERF_CORR_LMS_DCOFF_Q, cal->iq_q);  
+  bladerf_set_correction(handler->dev, BLADERF_MODULE_TX, BLADERF_CORR_LMS_DCOFF_Q, cal->iq_q);
 }
 
 void rf_blade_set_rx_cal(void *h, srslte_rf_cal_t *cal) {
@@ -404,7 +404,7 @@ void rf_blade_set_rx_cal(void *h, srslte_rf_cal_t *cal) {
   bladerf_set_correction(handler->dev, BLADERF_MODULE_RX, BLADERF_CORR_FPGA_PHASE, cal->dc_gain);
   bladerf_set_correction(handler->dev, BLADERF_MODULE_RX, BLADERF_CORR_FPGA_GAIN, cal->dc_phase);
   bladerf_set_correction(handler->dev, BLADERF_MODULE_RX, BLADERF_CORR_LMS_DCOFF_I, cal->iq_i);
-  bladerf_set_correction(handler->dev, BLADERF_MODULE_RX, BLADERF_CORR_LMS_DCOFF_Q, cal->iq_q);  
+  bladerf_set_correction(handler->dev, BLADERF_MODULE_RX, BLADERF_CORR_LMS_DCOFF_Q, cal->iq_q);
 }
 
 
@@ -426,11 +426,11 @@ static void secs_to_timestamps(uint32_t rate, time_t secs, double frac_secs, uin
   }
 }
 
-void rf_blade_get_time(void *h, time_t *secs, double *frac_secs) 
+void rf_blade_get_time(void *h, time_t *secs, double *frac_secs)
 {
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   struct bladerf_metadata meta;
-  
+
   int status = bladerf_get_timestamp(handler->dev, BLADERF_MODULE_RX, &meta.timestamp);
   if (status != 0) {
       fprintf(stderr, "Failed to get current RX timestamp: %s\n",
@@ -445,9 +445,9 @@ int rf_blade_recv_with_time_multi(void *h,
                                   uint32_t nsamples,
                                   bool blocking,
                                   time_t *secs,
-                                  double *frac_secs) 
+                                  double *frac_secs)
 {
-  return rf_blade_recv_with_time(h, *data, nsamples, blocking, secs, frac_secs);  
+  return rf_blade_recv_with_time(h, *data, nsamples, blocking, secs, frac_secs);
 }
 
 int rf_blade_recv_with_time(void *h,
@@ -455,15 +455,15 @@ int rf_blade_recv_with_time(void *h,
                     uint32_t nsamples,
                     bool blocking,
                     time_t *secs,
-                    double *frac_secs) 
+                    double *frac_secs)
 {
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   struct bladerf_metadata meta;
-  int status; 
-  
+  int status;
+
   memset(&meta, 0, sizeof(meta));
   meta.flags = BLADERF_META_FLAG_RX_NOW;
-  
+
   if (2*nsamples > CONVERT_BUFFER_SIZE) {
     fprintf(stderr, "RX failed: nsamples exceeds buffer size (%d>%d)\n", nsamples, CONVERT_BUFFER_SIZE);
     return -1;
@@ -474,27 +474,27 @@ int rf_blade_recv_with_time(void *h,
     return -1;
   } else if (meta.status & BLADERF_META_STATUS_OVERRUN) {
     if (blade_error_handler) {
-      srslte_rf_error_t error; 
+      srslte_rf_error_t error;
       error.opt = meta.actual_count;
       error.type = SRSLTE_RF_ERROR_OVERFLOW;
       blade_error_handler(error);
-    } else {     
+    } else {
       fprintf(stderr, "Overrun detected in scheduled RX. "
             "%u valid samples were read.\n\n", meta.actual_count);
     }
   }
-  
+
   timestamp_to_secs(handler->rx_rate, meta.timestamp, secs, frac_secs);
   srslte_vec_convert_if(handler->rx_buffer, 2048, data, 2*nsamples);
-  
+
   return nsamples;
 }
-                   
+
 int rf_blade_send_timed_multi(void *h,
                      void *data[4],
                      int nsamples,
                      time_t secs,
-                     double frac_secs,                      
+                     double frac_secs,
                      bool has_time_spec,
                      bool blocking,
                      bool is_start_of_burst,
@@ -508,27 +508,27 @@ int rf_blade_send_timed(void *h,
                      void *data,
                      int nsamples,
                      time_t secs,
-                     double frac_secs,                      
+                     double frac_secs,
                      bool has_time_spec,
                      bool blocking,
                      bool is_start_of_burst,
-                     bool is_end_of_burst) 
+                     bool is_end_of_burst)
 {
   rf_blade_handler_t *handler = (rf_blade_handler_t*) h;
   struct bladerf_metadata meta;
-  int status; 
-  
+  int status;
+
   if (!handler->tx_stream_enabled) {
     rf_blade_start_tx_stream(h);
   }
-  
+
   if (2*nsamples > CONVERT_BUFFER_SIZE) {
     fprintf(stderr, "TX failed: nsamples exceeds buffer size (%d>%d)\n", nsamples, CONVERT_BUFFER_SIZE);
     return -1;
   }
 
   srslte_vec_convert_fi(data, 2048, handler->tx_buffer, 2*nsamples);
-  
+
   memset(&meta, 0, sizeof(meta));
   if (is_start_of_burst) {
     if (has_time_spec) {
@@ -541,16 +541,16 @@ int rf_blade_send_timed(void *h,
   if (is_end_of_burst) {
     meta.flags |= BLADERF_META_FLAG_TX_BURST_END;
   }
-  srslte_rf_error_t error; 
+  srslte_rf_error_t error;
   bzero(&error, sizeof(srslte_rf_error_t));
-  
+
   status = bladerf_sync_tx(handler->dev, handler->tx_buffer, nsamples, &meta, 2000);
   if (status == BLADERF_ERR_TIME_PAST) {
     if (blade_error_handler) {
       error.type = SRSLTE_RF_ERROR_LATE;
       blade_error_handler(error);
     } else {
-      fprintf(stderr, "TX failed: %s\n", bladerf_strerror(status));      
+      fprintf(stderr, "TX failed: %s\n", bladerf_strerror(status));
     }
   } else if (status) {
     fprintf(stderr, "TX failed: %s\n", bladerf_strerror(status));
@@ -563,7 +563,7 @@ int rf_blade_send_timed(void *h,
       fprintf(stderr, "TX warning: underflow detected.\n");
     }
   }
-  
+
   return nsamples;
 }
